Fix cros thread input crashes
The key and text events from the java side were calling uno code from a
thread that was not the draw thread. This caused seg faults. I have now
added a mutex queue so events can be fed from the java thread to the
draw thread. This has meant adding the cross thread event struct which
is ugly but will probably be refactored soon.